+++ crates/biome_service/src/configuration/parse/json/rules.rs
@@ -2030,6 +2030,7 @@ impl VisitNode<JsonLanguage> for Nursery {
                 "noFallthroughSwitchClause",
                 "noGlobalIsFinite",
                 "noGlobalIsNan",
+                "noMisleadingInstantiator",
                 "noVoid",
                 "useArrowFunction",
                 "useCollapsedElseIf",
@@ -2210,6 +2211,29 @@ impl VisitNode<JsonLanguage> for Nursery {
                         diagnostics,
                     )?;
                     self.no_global_is_nan = Some(rule_configuration);
+                }
+                _ => {
+                    diagnostics.push(DeserializationDiagnostic::new_incorrect_type(
+                        "object or string",
+                        value.range(),
+                    ));
+                }
+            },
+            "noMisleadingInstantiator" => match value {
+                AnyJsonValue::JsonStringValue(_) => {
+                    let mut configuration = RuleConfiguration::default();
+                    self.map_to_known_string(&value, name_text, &mut configuration, diagnostics)?;
+                    self.no_misleading_instantiator = Some(configuration);
+                }
+                AnyJsonValue::JsonObjectValue(_) => {
+                    let mut rule_configuration = RuleConfiguration::default();
+                    rule_configuration.map_rule_configuration(
+                        &value,
+                        name_text,
+                        "noMisleadingInstantiator",
+                        diagnostics,
+                    )?;
+                    self.no_misleading_instantiator = Some(rule_configuration);
                 }
                 _ => {
                     diagnostics.push(DeserializationDiagnostic::new_incorrect_type(