+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Read our [guidelines for writing a good changelog entry](https://github.com/biom
 ### Formatter
 ### JavaScript APIs
 ### Linter
+
+#### New features
+
+- Add [noMisleadingInstantiator](https://biomejs.dev/linter/rules/no-mileading-instantiator) rule. The rule reports the misleading use of the `new` and `constructor` methods. Contributed by [@unvalley](https://github.com/unvalley)
+
 ### Parser
 ### VSCode